Dra ut i komponent
